Module 1

+++ tests/check_bucket.py
@@ -2,7 +2,7 @@ import numpy as np
 
 edges = np.fromfile(
     "data/buckets/bucket_001.bin",
-    dtype=np.uint32
+    dtype=np.uint64
 )
 
 edges = edges.reshape(-1, 2)
